docs(Table): fixed table crashing story and updated to new way of writing
