+++ src/components/Table/__stories__/Template.stories.tsx
@@ -1,0 +1,6 @@
+import { ComponentStory } from '@storybook/react'
+import Table from '..'
+
+export const Template: ComponentStory<typeof Table> = ({ ...props }) => (
+  <Table {...props} />
+)
