+++ src/components/Table/__stories__/index.stories.tsx
@@ -1,6 +1,5 @@
-import { Meta, Story } from '@storybook/react'
+import { Meta } from '@storybook/react'
 import Table, { Body, BodyCell, Head, HeadCell, Row } from '..'
-import Button from '../../Button'
 
 export default {
   component: Table,
@@ -15,74 +14,6 @@ export default {
   title: 'Components/Data Display/Table',
 } as Meta
 
-const Template: Story = args => <Table {...args} />
-
-Template.args = {
-  children: [
-    <Table.Head>
-      <Table.Row>
-        <Table.HeadCell>ID</Table.HeadCell>
-      </Table.Row>
-    </Table.Head>,
-    <Table.Body>
-      <Table.Row>
-        <Table.BodyCell>165</Table.BodyCell>
-      </Table.Row>
-    </Table.Body>,
-  ],
-}
-
-export const Default = Template.bind({})
-
-Default.args = {
-  ...Template.args,
-}
-
-export const Hoverable = Template.bind({})
-
-Hoverable.args = {
-  children: [
-    <Table.Head>
-      <Table.Row>
-        <Table.HeadCell>ID</Table.HeadCell>
-        <Table.HeadCell>Date</Table.HeadCell>
-        <Table.HeadCell>Action</Table.HeadCell>
-      </Table.Row>
-    </Table.Head>,
-    <Table.Body>
-      <Table.Row>
-        <Table.BodyCell>999</Table.BodyCell>
-        <Table.BodyCell>Just now</Table.BodyCell>
-        <Table.BodyCell>
-          <Button size="small" icon="eye" data-visibility="hover" />
-        </Table.BodyCell>
-      </Table.Row>
-      <Table.Row>
-        <Table.BodyCell>165</Table.BodyCell>
-        <Table.BodyCell>Yesterday</Table.BodyCell>
-        <Table.BodyCell>
-          <Button size="small" icon="eye" data-visibility="hover" />
-        </Table.BodyCell>
-      </Table.Row>
-    </Table.Body>,
-  ],
-}
-
-export const Loader = Template.bind({})
-
-Loader.args = {
-  children: [
-    <Table.Head>
-      <Table.Row>
-        <Table.HeadCell>ID</Table.HeadCell>
-        <Table.HeadCell>Date</Table.HeadCell>
-      </Table.Row>
-    </Table.Head>,
-    <Table.Body loading colSpan={2}>
-      <Table.Row>
-        <Table.BodyCell>999</Table.BodyCell>
-        <Table.BodyCell>Just now</Table.BodyCell>
-      </Table.Row>
-    </Table.Body>,
-  ],
-}
+export { Playground } from './Playground.stories'
+export { Hoverable } from './Hoverable.stories'
+export { Loader } from './Loader.stories'
